refactor(string-calculator): 변수명 수정 및 메소드 분리
