+++ src/main/java/calculator/StringSplitter.java
