--- src/test/java/calculator/ZeroAndPositiveNumber.java
@@ -1,18 +1,0 @@
-package calculator;
-
-public class ZeroAndPositiveNumber {
-    
-    private final int number;
-    
-    public ZeroAndPositiveNumber(String token) {
-        number = Integer.parseInt(token);
-        
-        if (number < 0) {
-            throw new RuntimeException();
-        }
-    }
-    
-    public int getValue() {
-        return number;
-    }
-}
